add dependency injection for UserConnector in LoginActivity
