+++ app/src/main/java/ch/epfl/sdp/dagger/modules/RepoModule.kt
@@ -1,14 +1,14 @@
-package ch.epfl.sdp.lobby
+package ch.epfl.sdp.dagger.modules
 
-import ch.epfl.sdp.lobby.game.FakeMockGameLobbyRepo
 import ch.epfl.sdp.lobby.game.IGameLobbyRepository
+import ch.epfl.sdp.lobby.game.MockGameLobbyRepository
 import dagger.Module
 import dagger.Provides
 
 @Module
-class FakeRepoModule {
+class RepoModule {
     @Provides
     fun providesMockGameLobbyRepo(): IGameLobbyRepository {
-        return FakeMockGameLobbyRepo
+        return MockGameLobbyRepository
     }
 }
