+++ app/src/main/java/ch/epfl/sdp/lobby/GameCreationActivity.kt
@@ -3,6 +3,7 @@ package ch.epfl.sdp.lobby
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import ch.epfl.sdp.dagger.MyApplication
 import ch.epfl.sdp.databinding.ActivityGameCreationBinding
 import ch.epfl.sdp.lobby.game.IGameLobbyRepository
 import java.sql.Time
