+++ app/src/main/java/ch/epfl/sdp/lobby/game/FakeMockGameLobbyRepo.kt
@@ -17,13 +17,7 @@ object FakeMockGameLobbyRepo : IGameLobbyRepository {
             Participation(User("George Kittle", 85), false, "CAFE", PlayerFaction.PREDATOR),
             Participation(User("Nick Bosa", 97), false, "0A0A", PlayerFaction.PREDATOR),
             Participation(User("Richard Sherman", 25), false, "C0BA", PlayerFaction.PREDATOR),
-            Participation(User("Dummy User", 23), false, "AB00", PlayerFaction.PREY),
-            Participation(User("Hello World", 42), true, "C0B0", PlayerFaction.PREY),
-            Participation(User("Morgan Freeman", 91), false, "0BBB", PlayerFaction.PREDATOR),
-            Participation(User("Jack Sparrow", 7), true, "0AAC", PlayerFaction.PREY),
-            Participation(User("Britney Spears", 24), false, "AC00", PlayerFaction.PREDATOR),
-            Participation(User("Spiderman", 25), false, "A0AA", PlayerFaction.PREDATOR),
-            Participation(User("Neymar Jr", 26), false, "C000", PlayerFaction.PREDATOR))
+            Participation(User("Dummy User", 23), false, "AB00", PlayerFaction.PREY))
 
     override fun createGame(gameName: String, gameDuration: Time): Int {
         return 42
@@ -43,6 +37,7 @@ object FakeMockGameLobbyRepo : IGameLobbyRepository {
     }
 
     override fun getParticipations(cb : Callback<List<Participation>>) {
+        assert(false)
         cb(players)
     }
