+++ app/src/main/java/ch/epfl/sdp/lobby/game/GameLobbyActivity.kt
@@ -13,7 +13,9 @@ import ch.epfl.sdp.databinding.ActivityGameLobbyBinding
 import ch.epfl.sdp.game.PlayerFaction
 import ch.epfl.sdp.game.PredatorActivity
 import ch.epfl.sdp.game.PreyActivity
+import ch.epfl.sdp.dagger.MyApplication
 import ch.epfl.sdp.lobby.PlayerParametersFragment
+import javax.inject.Inject
 
 /**
  * Game Lobby Activity showing the list of players and game info
@@ -28,11 +30,13 @@ class GameLobbyActivity : AppCompatActivity(), SwipeRefreshLayout.OnRefreshListe
     private lateinit var mSwipeRefreshLayout: SwipeRefreshLayout
     private lateinit var rv: RecyclerView
     private var adminId = 0
-    private val repository = MockGameLobbyRepository
+    @Inject lateinit var repository : IGameLobbyRepository;
     private lateinit var gameLobbyBinding: ActivityGameLobbyBinding
     private var myFaction: PlayerFaction = PlayerFaction.PREDATOR
 
     override fun onCreate(savedInstanceState: Bundle?) {
+        (applicationContext as MyApplication).appComponent.inject(this)
+
         super.onCreate(savedInstanceState)
         gameLobbyBinding = ActivityGameLobbyBinding.inflate(layoutInflater)
         setContentView(gameLobbyBinding.root)
